feat(loans-page): Loans CRUD and frontend fixes

+++ apps/api/src/loan/loan.service.ts
@@ -7,7 +7,7 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { LoanEntity } from './entities/loan.entity';
 
 import { CreateLoanDto, LoanStatus } from '@repo/common';
-import { Repository } from 'typeorm';
+import { Repository, UpdateResult } from 'typeorm';
 import { BookService } from '../book/services/book.service';
 import { GenericService } from '../core/generic.service';
 
@@ -24,28 +24,53 @@ export class LoanService extends GenericService {
 	async create(data: CreateLoanDto) {
 		const book = await this.bookService.findOne(data.bookId);
 		if (!book) throw new NotFoundException('Book not found');
+		if (new Date(Date.now()) > data.returnDate)
+			throw new BadRequestException('Return date cannot be in the past');
 		if (book.availableQuantity < data.quantity)
 			throw new BadRequestException('Not enough books available');
 		const loan = this.loanRepository.create({
 			...data,
+			loanDate: new Date(Date.now()),
 			book: {
 				id: data.bookId,
 			},
 		});
-		return await this.loanRepository.manager.transaction(
+		await this.loanRepository.manager.transaction(
 			async (transactionalEntityManager) => {
-				await transactionalEntityManager.save(loan);
 				await this.bookService.removeStock(data.bookId, data.quantity);
-				return this.findById(loan[0].id);
+				await transactionalEntityManager.save(loan);
 			},
 		);
+		return loan;
 	}
-	async returnBook(loanId: number): Promise<void> {
-		const loan: LoanEntity = await this.findById(loanId);
+	async returnBook(loanId: number): Promise<UpdateResult> {
+		const loan = await this.loanRepository.findOne({
+			where: { id: loanId },
+			relations: ['book'],
+		});
+		if (!loan) throw new NotFoundException('Loan not found');
 		const book = await this.bookService.findById(loan.book.id);
-		book.availableQuantity += loan.quantity;
-		await this.bookService.updateStock(book.id, book.availableQuantity);
+		await this.bookService.updateStock(book.id, loan.quantity);
 		loan.status = LoanStatus.RETURNED;
-		await this.update(loanId, loan);
+		loan.returnDate = new Date(Date.now());
+		return await this.update(loanId, loan);
+	}
+
+	async paginatedLoans(page: number) {
+		const [data, total] = await this.loanRepository.findAndCount({
+			relations: ['book'],
+			order: { id: 'ASC' },
+			take: 6,
+			skip: (page - 1) * 6,
+		});
+		return {
+			data: data.map((loan) => ({
+				...loan,
+				book: loan.book.title,
+			})),
+			total,
+			current_page: page,
+			last_page: Math.ceil(total / 6),
+		};
 	}
 }
